// packages/solid-webgpu-gltf/src/generated/glTF.ts
export type GlTFChildOfRootProperty = GlTFProperty & {
  /**
   * The user-defined name of this object.
   */
  name?: string
  [k: string]: unknown
}
export type GlTFId = number
// ...
export type Material = GlTFChildOfRootProperty & {
  name?: unknown
  extensions?: unknown
  extras?: unknown
  /**
   * A set of parameter values that are used to define the metallic-roughness material model from Physically Based Rendering (PBR) methodology. When undefined, all the default values of `pbrMetallicRoughness` **MUST** apply.
   */
  pbrMetallicRoughness?: MaterialPBRMetallicRoughness
  /**
   * The tangent space normal texture.
   */
  normalTexture?: MaterialNormalTextureInfo
  /**
   * The occlusion texture.
   */
  occlusionTexture?: MaterialOcclusionTextureInfo
  /**
   * The emissive texture.
   */
  emissiveTexture?: TextureInfo
  /**
   * The factors for the emissive color of the material.
   *
   * @minItems 3
   * @maxItems 3
   */
  emissiveFactor?: number[]
  /**
   * The alpha rendering mode of the material.
   */
  alphaMode?: 'OPAQUE' | 'MASK' | 'BLEND' | string
  /**
   * The alpha cutoff value of the material.
   */
  alphaCutoff?: number
  /**
   * Specifies whether the material is double sided.
   */
  doubleSided?: boolean
  [k: string]: unknown
}
/**
 * A set of parameter values that are used to define the metallic-roughness material model from Physically-Based Rendering (PBR) methodology.
 */
export type MaterialPBRMetallicRoughness = GlTFProperty & {
  /**
   * The factors for the base color of the material.
   *
   * @minItems 4
   * @maxItems 4
   */
  baseColorFactor?: number[]
  /**
   * The base color texture.
   */
  baseColorTexture?: TextureInfo
  /**
   * The factor for the metalness of the material.
   */
  metallicFactor?: number
  /**
   * The factor for the roughness of the material.
   */
  roughnessFactor?: number
  /**
   * The metallic-roughness texture.
   */
  metallicRoughnessTexture?: TextureInfo
  extensions?: unknown
  extras?: unknown
  [k: string]: unknown
}
/**
 * Reference to a texture.
 */
export type TextureInfo = GlTFProperty & {
  /**
   * The index of the texture.
   */
  index: GlTFId
  /**
   * The set index of texture's TEXCOORD attribute used for texture coordinate mapping.
   */
  texCoord?: number
  extensions?: unknown
  extras?: unknown
  [k: string]: unknown
}
export type MaterialNormalTextureInfo = TextureInfo & {
  index?: unknown
  texCoord?: unknown
  /**
   * The scalar parameter applied to each normal vector of the normal texture.
   */
  scale?: number
  extensions?: unknown
  extras?: unknown
  [k: string]: unknown
}
export type MaterialOcclusionTextureInfo = TextureInfo & {
  index?: unknown
  texCoord?: unknown
  /**
   * A scalar multiplier controlling the amount of occlusion applied.
   */
  strength?: number
  extensions?: unknown
  extras?: unknown
  [k: string]: unknown
}
// ...
export interface GlTFProperty {
  extensions?: Extension
  extras?: Extras
  [k: string]: unknown
}